+++ AddOns/DBM-PvP/Battlegrounds/Ashran.lua
@@ -1,7 +1,7 @@
 local mod	= DBM:NewMod("z1191", "DBM-PvP") -- Added in WoD
 
 
-mod:SetRevision("20240505221847")
+mod:SetRevision("20240903194526")
 mod:SetZone(DBM_DISABLE_ZONE_DETECTION)
 mod:RegisterEvents(
 	"LOADING_SCREEN_DISABLED",
@@ -56,8 +56,8 @@ do
 		end
 		local cid = self:GetCIDFromGUID(UnitGUID("target") or "")
 		if cid == 81870 or cid == 82204 or cid == 183198 then -- Anenga (Alliance) | Atomik/Narduke (Horde)
-			local _, currency = GetCurrencyInfo(944) -- Artifact Fragment
-			if currency > 0 and GetNumGossipOptions() == 3 then -- If boss isn't already summoned
+			local currency = GetCurrencyInfo(944) -- Artifact Fragment
+			if currency and currency.quantity > 0 and GetNumGossipOptions() == 3 then -- If boss isn't already summoned
 				local gossipOptionID = self:GetGossipID()
 				if gossipOptionID then
 					self:SelectGossip(gossipOptionID)