+++ AddOns/DBM-PvP/CHANGELOG.md
@@ -1,6 +1,7 @@
 # <DBM Mod> PvP
 
-## [r179](https://github.com/DeadlyBossMods/DBM-PvP/tree/r179) (2024-07-22)
-[Full Changelog](https://github.com/DeadlyBossMods/DBM-PvP/compare/r178...r179) [Previous Releases](https://github.com/DeadlyBossMods/DBM-PvP/releases)
+## [r180](https://github.com/DeadlyBossMods/DBM-PvP/tree/r180) (2024-09-03)
+[Full Changelog](https://github.com/DeadlyBossMods/DBM-PvP/compare/r179...r180) [Previous Releases](https://github.com/DeadlyBossMods/DBM-PvP/releases)
 
-- Fix type error. Closes 161  
+- Fix ashran currency detection on gossip  
+- remove 10.2.7 and 11.0.0 game versions from tocs